Bug 2002522 - Part 2: Fix deprecated enum-enum conversion warnings for mips64. r=spidermonkey-reviewers,jandem
Add explicit casts to squelch deprecated-enum-enum-conversion warnings.

Also see bug 1791955.

Differential Revision: https://phabricator.services.mozilla.com/D274122

UltraBlame original commit: db0aa0e463754170425c3e74f2c262bfc96aaee5

+++ js/src/jit/mips-shared/Assembler-mips-shared.cpp
@@ -251,7 +251,7 @@ BufferOffset AssemblerMIPSShared::nopAlign(int alignment) {
 
 BufferOffset AssemblerMIPSShared::as_nop() {
   spew("nop");
-  return writeInst(op_special | ff_sll);
+  return writeInst(static_cast<uint32_t>(op_special) | ff_sll);
 }
